solved with horrific workaround.
Output stream is piped to a temp file, which will need to be
formalised: Name the temp file after the process id or something. I
don’t know yet. It needs to have its own folder to live in. This is a
retarded way of doing it, but it’s really no worse than python does
with input and output anyway.

Temp file is read and interpreted as the percentage to set for the
processing. This can be done in the same way as the original plan. Will
formalise in the morning.